Exclude nodes based on labels
This fits in with the idea that we are viewing worker nodes as just a pool
of resources and not by individual hostnames. This gives the flexibility for
users to isolate the tests to only one hardware model of workers as well (since
each model can be labelled with its model name).

Also there was a bug in previous code where only the first node in the list
was technically being excluded in `workload_args.excluded_node[0]`.

Signed-off-by: Sai Sindhur Malleni <smalleni@redhat.com>